abc 094 c

+++ etc/tmp.cpp
@@ -13,7 +13,7 @@ typedef vector<VI> VVI;
 
 int main() {
     ios::sync_with_stdio(false); cin.tie(0);
-    int N;
+    ll N;
     cin >> N;
     
     OUT("");
